requirements.txt
pytest==6.0.1
selenium==3.141.0
allure-pytest==2.8.16
allure-python-commons==2.8.16
